requirements_videoanalytics.txt
streamlit
torch
torchvision
torchaudio
opencv-python
pillow
numpy